consumer/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
joblib==1.1.0
kafka-python==2.0.2
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
py4j==0.10.9.3
pymongo==4.1.1
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
spark-nlp==3.4.3
threadpoolctl==3.1.0
urllib3==1.26.9
